moved Input component into Header folder
